backend/src/requirements.txt
flask==1.1.2
flask-restful==0.3.8
docopt==0.6.2
flask-sqlalchemy==2.4.4
sqlalchemy-utils==0.36.8
sqlalchemy==1.3.22
Flask-Script==2.0.6
Flask-Migrate==2.6.0
werkzeug==1.0.1
pyjwt==2.0.1
flask-httpauth==4.2.0
flask-api==2.0
pyinstaller==4.2
